ui(desktop): convert Settings from modal to full-screen route
The reference is a Codex-style settings UI with a fixed left rail
("Back to app" + lucide-icon nav entries) and a generously-padded
right column hosting a page heading, section labels, and grouped
setting cards. This commit migrates the existing dialog into that
shape without restructuring AppShell's render tree.

## New primitives (desktop/src/components/settings/)

- **SettingsScreen** — full-height shell. 240px nav rail with a "Back
  to app" entry at top (chevron-left + muted text), section list
  below (lucide icon + label, selected state = bg-accent + medium
  weight, idle = muted-foreground). Right column is a single
  scrollable region, owns its own scrollbar gutter, and respects the
  macOS stoplight pad via the existing `useStoplightCompensation`
  hook.

- **SettingsPage** — inner content wrapper. Centers at `max-w-3xl`,
  generous `px-10 py-12`, page-level `<h1>` (text-2xl semibold) +
  optional subtitle, then a `gap-10` stack for sections.

- **SettingsScreenRoot** — the entry-point that AppShell mounts. Owns
  the section-id state machine, the diagnostics export flow, the
  app-update channel state, and renders the active panel body
  (Account / Billing / AI / Integrations / Submissions / General /
  About) into a `SettingsPage`. All panel bodies are lifted from the
  old SettingsDialog with their behavior unchanged.

## Refreshed primitives

- **SettingsCard**: `shadow-md` → single hairline `border border-border`.
  Matches the reference's quieter chrome (no card lift) — the page
  itself is the visual frame; cards are just groupings.
- **SettingsRow**: `py-3` → `py-3.5`. Half-step vertical bump for the
  reference's roomier rhythm.
- **SettingsSection** unchanged.

## AppShell wiring

The old `SettingsDialog` was a Base UI Dialog mounted at the bottom
of `<main>`. It's replaced with a conditional full-cover overlay
(`absolute inset-0 z-40 bg-background`) that hosts `SettingsScreenRoot`
when `settingsDialogOpen` is true. Keeps the same open/close state
machine + IPC `onOpenSettingsPane` channel; the `onClose` handler
becomes `onBackToApp` semantically.

State variable names (`settingsDialogOpen`, `settingsDialogSection`)
keep their dialog-era spelling so the diff stays focused on the
visual + structural change. Renaming is left as follow-up cleanup.

## Section ordering

Reordered the rail to match the reference: General first (most-common
destination), then Account, Billing, AI, Integrations, Submissions,
About. Section ids stay the legacy `UiSettingsPaneSection` strings
(`"settings"` etc.) so the IPC `onOpenSettingsPane(section)` channel
is unaffected; only labels + order changed.

Each page also gets a one-line subtitle pulled from a small
`pageDescription()` switch, matching the reference's "Choose how
much technical detail …" pattern under the page heading.

## Verification

- `npm run desktop:typecheck` — clean
- `npm --prefix desktop run build:renderer` — clean
- `node --test` — same baseline as upstream/main (45 pass / 1
  pre-existing fail on the `updateNotification` brittle assertion).
  Updated AppShell.test.mjs to assert on `<SettingsScreenRoot>`
  instead of the deleted `<SettingsDialog>` JSX shape.
