+++ desktop/src/components/settings/SettingsCard.tsx
@@ -12,19 +12,19 @@ interface SettingsCardProps {
  *
  * A boxed surface for SettingsRow / SettingsToggle / SettingsMenuSelectRow.
  * Direct children automatically get a top border so successive rows share
- * a hairline divider (matches craft-agents-oss's settings card pattern).
+ * a hairline divider.
  *
  * Style notes:
- *  - shadow-md gives the unified hairline ring + subtle lift (one of our
- *    new design tokens — auto-tunes per mode via --hairline-alpha).
- *  - rounded-xl matches the SettingsDialog's other cards.
- *  - overflow-hidden so the inner row borders don't peek past the corners.
+ *  - Single hairline border around the card; no shadow. Matches the
+ *    full-screen settings reference's quieter chrome.
+ *  - rounded-xl + overflow-hidden so the inner row borders don't peek
+ *    past the corners.
  */
 export function SettingsCard({ children, className }: SettingsCardProps) {
   return (
     <div
       className={cn(
-        "overflow-hidden rounded-xl bg-card shadow-md [&>*+*]:border-t [&>*+*]:border-border",
+        "overflow-hidden rounded-xl border border-border bg-card [&>*+*]:border-t [&>*+*]:border-border",
         className,
       )}
     >
