+++ desktop/src/components/settings/index.ts
@@ -3,7 +3,13 @@ export {
   SettingsMenuSelectRow,
   type SettingsMenuOption,
 } from "./SettingsMenuSelectRow";
+export { SettingsPage, type SettingsPageProps } from "./SettingsPage";
 export { SettingsRow } from "./SettingsRow";
+export {
+  SettingsScreen,
+  type SettingsScreenNavEntry,
+  type SettingsScreenProps,
+} from "./SettingsScreen";
 export { SettingsSection } from "./SettingsSection";
 export {
   SettingsStatusBadge,
